# api/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Intelligence and Action Engine API
  version: 0.1.0
  description:
    Initial contract for authenticated workspaces, knowledge, conversations, memory, tools, approvals, audit, and
    evaluations.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /v1/workspaces/{workspace_id}/documents:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags:
        - Documents
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: project_id
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DocumentVersionStatus'
      responses:
        '200':
          description: Documents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'
  /v1/workspaces/{workspace_id}/documents/{document_id}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags:
        - Documents
      operationId: getDocument
      responses:
        '200':
          description: Document and version summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Documents
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Deletion workflow accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAccepted'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/workspaces/{workspace_id}/documents/{document_id}/ingestion-jobs:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DocumentId'
    post:
      tags:
        - Documents
      operationId: createIngestionJob
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Ingestion/retry accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
  /v1/workspaces/{workspace_id}/ingestion-jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - name: job_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags:
        - Documents
      operationId: getIngestionJob
      responses:
        '200':
          description: Ingestion job state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
...
  responses:
...
    NotFound:
      description: Resource not found or not visible to this principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              $ref: '#/components/schemas/Uuid'
            details:
              type: object
              additionalProperties: true
...
    Sensitivity:
      type: string
      enum:
        - PUBLIC
        - INTERNAL
        - CONFIDENTIAL
        - HIGHLY_CONFIDENTIAL
        - REGULATED
        - PROHIBITED
...
    DocumentVersionStatus:
      type: string
      enum:
        - PENDING_UPLOAD
        - UPLOADED
        - QUARANTINED
        - INGESTING
        - READY
        - FAILED
        - SUPERSEDED
        - DELETED
    Document:
      type: object
      required:
        - id
        - workspace_id
        - title
        - sensitivity
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        workspace_id:
          $ref: '#/components/schemas/Uuid'
        project_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        title:
          type: string
        sensitivity:
          $ref: '#/components/schemas/Sensitivity'
        current_version:
          oneOf:
            - $ref: '#/components/schemas/DocumentVersion'
            - type: 'null'
        created_at:
          $ref: '#/components/schemas/Timestamp'
    DocumentVersion:
      type: object
      required:
        - id
        - document_id
        - version_number
        - status
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        document_id:
          $ref: '#/components/schemas/Uuid'
        version_number:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/DocumentVersionStatus'
        is_current:
          type: boolean
        checksum_sha256:
          type: string
        created_at:
          $ref: '#/components/schemas/Timestamp'
    DocumentPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        next_cursor:
          type:
            - string
            - 'null'
    IngestionJob:
      type: object
      required:
        - id
        - document_version_id
        - status
        - attempt
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        document_version_id:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - RETRY_WAIT
            - SUCCEEDED
            - FAILED_FINAL
            - CANCELLED
        stage:
          type:
            - string
            - 'null'
        attempt:
          type: integer
          minimum: 0
        error_code:
          type:
            - string
            - 'null'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    OperationAccepted:
      type: object
      required:
        - operation_id
        - status
      properties:
        operation_id:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum:
            - ACCEPTED